requirements.txt
requests
fastapi
uvicorn
gitpython
